Refactor tags.service to autotags.service!
